+++ pypechain/utilities/format.py
@@ -6,7 +6,8 @@ import black
 
 
 def avoid_python_keywords(name: str) -> str:
-    """Make sure the variable name is not a reserved Python word.  If it is, prepend with an underscore.
+    """Make sure the variable name is not a reserved Python word.  If it is, prepend with an
+    underscore.
 
     Arguments
     ---------
