+++ pyproject.toml
@@ -32,3 +32,9 @@ build = ["build", "setuptools", "wheel", "twine"]
 [build-system]
 requires = ["flit_core>=3.2"]
 build-backend = "flit_core.buildapi"
+
+[tool.pylint.'MASTER']
+extension-pkg-allow-list = ["black"]
+
+[tool.pylint.'MESSAGES CONTROL']
+disable = ["fixme"]
